ref: remove generics from some classes, use any in mixins
Essentially the generics get in the way of actually using the basic type
stubs. We can add them back later once this gets put into use and we can
see how beneficial they are.

Removed the `Request` types and the `AbstractBaseUser` types for a few
methods in favor of `Any` so subclasses can more freely replace those
with a more precise type without having to deal with the override rules.